+++ src/components/Modal/Modal.js
@@ -7,7 +7,7 @@ import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import LinearProgress from '@material-ui/core/LinearProgress';
 
-export default function FormDialog({htmlmarkup=null, modal, error, setModal, title, description, children, primaryButton, fullWidth=true, fullScreen=false, handlePrimaryAction, secondaryButton, handleSecondaryAction, contentContainerStyle={}, loading}) {
+export default function FormDialog({htmlmarkup=null, actionContainerStyle={}, modal, error, setModal, title, description, children, primaryButton, fullWidth=true, fullScreen=false, handlePrimaryAction, secondaryButton, handleSecondaryAction, contentContainerStyle={}, loading}) {
   const [open, setOpen] = React.useState(false);
   useEffect(() => {
     if (modal.hasOwnProperty('show')) {
@@ -35,7 +35,7 @@ export default function FormDialog({htmlmarkup=null, modal, error, setModal, tit
           </DialogContentText>}
             {children}
         </DialogContent>
-        <DialogActions>
+        <DialogActions style={actionContainerStyle}>
           <Button onClick={handleClose} color="primary">
             {secondaryButton}
           </Button>